Hack to fix Humidity and gas resisitance readouts
By runing bme280 lib init and config on bme680, the humidity and
gas resistance start reading out plausible values, instead of
constantly reading the max.

This may be from an extra soft reset before reading data. More
work must be performed to check which registers are being written
by the BME280 library and how those overlap with the registers
on the BME680.