Fixed race condition in ESP-AT library
When the send command was sent to the WiFi module, another
thread could try to send a command between the send command and
the sending of the data.

To allow the mutex protection to remain in place for the
esp_at_send_data() command, mutexes were changed to recursive
mutexes and entry and exits were added to the send data command.